+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 9.3.0
+
+### Inspector
+
+- Inspector: CLI integration in Playground - by [ryantrem](https://github.com/ryantrem) ([#18271](https://github.com/BabylonJS/Babylon.js/pull/18271))
+
+### Playground
+
+- Inspector: CLI integration in Playground - by [ryantrem](https://github.com/ryantrem) ([#18271](https://github.com/BabylonJS/Babylon.js/pull/18271))
+
 ## 9.2.2
 
 ### Addons